REDDEV-510 Refactor to add form for adding counts
Refactored to support a form for adding report counts.

settings.php no longer persists sample configuration if not present.

Extracted ReportSummaryForm component.

Default view is an empty page with "Create a Report" button. Clicking
the button opens a form for entering a title and a report. Upon save the
report configuration is saved by settings.php. Either saves new
configuration to the database, or appends to the existing config.

Currently the form remains on the screen and multiple summaries can be
added.

Removed sample report config and updated documentation and tests.