REDDEV-595 permissions
This branch includes work to restrict access for modifying counts.

User based settings permissions are disabled when saving module settings. This allows
all users to save settings if they have 'Add/Edit/Organize' rights, or are a super user.

Form and summary controls are hidden if the user does not have 'Add/Edit/Organize' rights, or is not a super user.

redcap_module_link_check_display is overridden so that everyone with project access can use the module.